+++ Sources/TuistKit/Services/ListService.swift
@@ -7,6 +7,10 @@ import TuistSupport
 class ListService {
     private let templatesDirectoryLocator: TemplatesDirectoryLocating
     private let templateLoader: TemplateLoading
+    private let textTable = TextTable<PrintableTemplate>{[
+        TextTable.Column(title: "Name", value: $0.name),
+        TextTable.Column(title: "Description", value: $0.description),
+    ]}
 
     init(templatesDirectoryLocator: TemplatesDirectoryLocating = TemplatesDirectoryLocator(),
          templateLoader: TemplateLoading = TemplateLoader()) {
@@ -18,11 +22,13 @@ class ListService {
         let path = self.path(path)
 
         let templateDirectories = try templatesDirectoryLocator.templateDirectories(at: path)
-
-        try templateDirectories.forEach {
-            let template = try templateLoader.loadTemplate(at: $0)
-            logger.info("\($0.basename): \(template.description)")
+        let templates: [PrintableTemplate] = try templateDirectories.map { path in
+            let template = try templateLoader.loadTemplate(at: path)
+            return PrintableTemplate(name: path.basename, description: template.description)
         }
+        
+        let renderedTable = textTable.render(templates)
+        logger.info("\(renderedTable)")
     }
 
     // MARK: - Helpers
@@ -34,4 +40,9 @@ class ListService {
             return FileHandler.shared.currentPath
         }
     }
+}
+
+private struct PrintableTemplate {
+    let name: String
+    let description: String
 }